+++ my_music_code/lib/Globals/navigator_page.dart
@@ -2,9 +2,9 @@
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:my_music_code/Album/album_page.dart';
 import 'package:my_music_code/Feed/feed_page.dart';
 import 'package:my_music_code/Globals/mini_player.dart';
+import 'package:my_music_code/Models/album_model.dart';
 import 'package:my_music_code/Models/music_model.dart';
 import 'package:my_music_code/MyPlaylists/user_page_of_playlists.dart';
 import 'package:my_music_code/Search/search_page.dart';
@@ -31,6 +31,7 @@ class _NavigatorPageState extends State<NavigatorPage> {
       universal.user = widget.user;
       universal.spotifyApi = widget.spotifyApi;
     });
+    // ignore: avoid_print
     print(universal.userModel.username);
     if (universal.userModel.username.isNotEmpty) {
       widget.user.updateProfile(displayName: universal.userModel.username);
