+++ my_music_code/lib/SpotifyApi/get_data.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
-
-import 'package:my_music_code/Album/album_page.dart';
+import 'package:my_music_code/Models/album_model.dart';
 import 'package:my_music_code/Models/music_model.dart';
 import 'package:spotify/spotify.dart';
 import 'package:my_music_code/universal.dart' as universal;
